Provide way to write request body
It is sometimes convenient to write to the request body.  One example is sending a multipart request body with large files stored on disk. 

It's possible to get a writer on the request body using io.Pipe and a goroutine, but it would be nice of the client handle these details or if the client plumbed a writer more directly to the network.